requirements.txt
Django==4.2.3
phonenumbers==8.13.17
Pillow==9.5.0
Pillow==10.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
dj-database-url==2.0.0